Reorganized code structure
