+++ apps/web/components/logo.tsx
@@ -1,4 +1,4 @@
-function Icon() {
+export function Logo() {
   return (
     <svg
       xmlns="http://www.w3.org/2000/svg"
@@ -23,5 +23,3 @@ function Icon() {
     </svg>
   )
 }
-
-export default Icon
