+++ apps/web/utils/firebase-config.ts
@@ -1,4 +1,4 @@
-const config = {
+export const config = {
   apiKey: process.env.NEXT_PUBLIC_FIREBASE_KEY,
   authDomain: `${process.env.NEXT_PUBLIC_FIREBASE_PID}.firebaseapp.com`,
   databaseURL: `https://${process.env.NEXT_PUBLIC_FIREBASE_PID}.firebaseio.com`,
@@ -8,5 +8,3 @@ const config = {
   appId: process.env.NEXT_PUBLIC_FIREBASE_APP_ID,
   measurementId: process.env.NEXT_PUBLIC_FIREBASE_MEASURE_ID,
 }
-
-export default config
